+++ packages/vendure-plugin-shipmate/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - Add `ShipmatePluginConfig.shouldSendOrder` (#498)
 
+# 1.2.4 (2024-08-20)
+
+- Divide by 100 when passing `OrderLine.proratedUnitPriceWithTax` to `Items.item_value`
+
 # 1.2.3 (2024-08-20)
 
 - Log warning when failed to cancel a shipment in Shipmate